requirements.txt
# tm-scan requirements
# Python 3.9+

pyyaml>=6.0
requests>=2.31.0
markdown>=3.6
weasyprint>=62.3
